[compat] fixed sample serializer
