+++ OpenRTM_aist/examples/Serializer/rtc_original_serializer.conf
@@ -4,7 +4,7 @@ logger.enable: YES
 logger.file_name: stdout
 logger.log_level: ERROR
 manager.modules.load_path: .
-manager.modules.preload: ConsoleOutDouble, TestSerializer1
+manager.modules.preload: ConsoleOutDouble, TestSerializer
 manager.components.precreate: ConsoleOutDouble
-#manager.components.preconnect: ConsoleOutDouble0.in?port=ConsoleInShort0.out&outport.marshaling_type=test1:dummydata&inport.marshaling_type=test1:dummydata
+#manager.components.preconnect: ConsoleOutDouble0.in?port=ConsoleInShort0.out&outport.marshaling_type=test:dummydata&inport.marshaling_type=test:dummydata
 #manager.components.preactivation: ConsoleOutDouble0, ConsoleInShort0
